Thêm tính năng nhập điểm cho sinh viên trong lớp học và cập nhật database

+++ QuanLyDiemSinhVien/DTO/DiemHocKy.cs
@@ -16,7 +16,7 @@ namespace DTO
     {
         public string MaHocKy { get; set; }
         public string MSSV { get; set; }
-        public Nullable<int> Diem { get; set; }
+        public Nullable<double> Diem { get; set; }
     
         public virtual HocKy HocKy { get; set; }
         public virtual SinhVien SinhVien { get; set; }
